divided HttpRequest.cpp; added check for receive bytes with content-length
