+++ conf/default.conf
@@ -38,7 +38,7 @@ http
             root    website
         }
 
-        location /campus19.be
+        location /campus19
         {
             return 301 http://campus19.be
         }
