+++ srcs/http_requests/HttpRequest.cpp
@@ -6,7 +6,7 @@
 /*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/29 02:18:22 by jdagoy            #+#    #+#             */
-/*   Updated: 2024/09/25 11:47:31 by jdagoy           ###   ########.fr       */
+/*   Updated: 2024/09/25 12:47:55 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,126 +69,7 @@ void HttpRequest::parseHttpRequest()
     }
 }
 
-std::string    HttpRequest::extract_token(const std::string &line, size_t &pos, char del)
-{
-    size_t  end = line.find(del, pos);
-    if (end == std::string::npos)
-        end = line.length();
-    std::string token = line.substr(pos, end - pos);
-    pos = end + 1;
-    return (token);
-}
-
-
-
-void    HttpRequest::parseRequestLine(const std::string &line)
-{
-    size_t  pos = 0;
-    
-    if (_request.getMethod().empty())
-    {
-        std::string method = extract_token(line, pos, ' '); 
-        if (method.empty())
-        {
-            setStatusCode(BAD_REQUEST);
-            return ;
-        }
-        trimWhitespaces(method);
-        _request.setMethod(method);
-    }
-
-    if (_request.getUri().empty())
-    {
-        std::string uri = extract_token(line, pos, ' ');
-        if (uri.empty())
-        {
-            setStatusCode(BAD_REQUEST);
-            return ;
-        }
-        if (uri.length() > MAX_URI_LENGTH)
-        {
-            setStatusCode(URI_TOO_LONG);
-            return ;
-        }
-        trimWhitespaces(uri);
-        _request.setUri(uri);
-    }
-
-    if (_request.getVersion().empty())
-    {
-        std::string version = extract_token(line, pos, ' ');
-        if (version.empty())
-        {
-            setStatusCode(BAD_REQUEST);
-            return ;
-        }
-        trimWhitespaces(version);
-        _request.setVersion(version);
-    }
-}
-
-bool    HttpRequest::isValidFieldName(const std::string &line)
-{
-    if (line.empty())
-        return (false);
-    for (size_t i = 0; i < line.size(); i++)
-        if (!std::isalnum(line[i]) && line[i] != '-')
-            return (false);
-    return (true);
-}
-
-bool    HttpRequest::isValidFieldValue(const std::string &line)
-{
-    if (line.empty())
-        return (false);
-    for (size_t i = 0; i < line.size(); i++)
-    {
-        if (!std::isprint(line[i]) && !std::isspace(line[i]) )
-            return (false);
-    }
-    return (true);
-}
-
-void HttpRequest::parseRequestHeaders(const std::string &line)
-{
-    size_t  colonPos = line.find (':');
-    if (colonPos == std::string::npos)
-    {
-        setStatusCode(BAD_REQUEST);
-        return ;
-    }
-
-    std::string fieldName = line.substr(0, colonPos);
-    std::string fieldValue = line.substr(colonPos + 1);
-    
-    fieldName = toLower(fieldName);
-    // for (size_t i = 0; i < fieldName.size(); i++)
-    //     fieldName[i] = std::tolower(fieldName[i]);
-    if (!isValidFieldName(fieldName) || !isValidFieldValue(fieldValue))
-    {
-        setStatusCode(BAD_REQUEST);
-        return ;
-    }    
-    trimWhitespaces(fieldValue);
-    if (fieldValue.length() > MAX_HEADER_LENGTH)
-    {
-        setStatusCode(REQUEST_HEADER_TOO_LARGE);
-        return ;
-    }
-    
-    std::stringstream ss(fieldValue);
-    std::string value;
-    std::vector<std::string> values;
-    while (std::getline(ss, value, ','))
-    {
-        trimWhitespaces(value);
-        values.push_back(value);
-    }
-    _headers[fieldName] = values;
-}
-
-
-size_t  getContentLengthBuffer(const std::string &header)
+size_t getContentLengthBuffer(const std::string &header)
 {
     std::string lowerHeader = toLower(header);
     size_t  pos = lowerHeader.find("content-length:");
@@ -201,7 +82,7 @@ size_t  getContentLengthBuffer(const std::string &header)
 }
 
 
-void    HttpRequest::requestToBuffer()
+void HttpRequest::requestToBuffer()
 {
     std::vector<unsigned char> buffer(1024);
     
@@ -216,7 +97,7 @@ void    HttpRequest::requestToBuffer()
         _buffer.insert(_buffer.end(), buffer.begin(), buffer.begin() + bytesRead);
     }
     ssize_t contentLength = getContentLengthBuffer(std::string(buffer.begin(), buffer.end()));;
-    if (contentLength == 0)
+    if (static_cast<ssize_t>(_buffer.size()) >= contentLength)
         return ;
     contentLength -= bytesRead;
     while (contentLength)
@@ -231,8 +112,8 @@ void    HttpRequest::requestToBuffer()
         {
             _buffer.insert(_buffer.end(), buffer.begin(), buffer.begin() + bytesRead);
         }
-        if (bytesRead > contentLength)
-            contentLength -= contentLength;
+        if (bytesRead >= contentLength)
+            contentLength = 0;
         else
             contentLength -= bytesRead;
     }    
@@ -249,7 +130,7 @@ std::vector<unsigned char>::iterator HttpRequest::findBufferCRLF()
     return (_buffer.end());
 }
 
-std::string    HttpRequest::getLineAndPopFromBuffer()
+std::string HttpRequest::getLineAndPopFromBuffer()
 {
     std::string line;
     
@@ -271,7 +152,7 @@ std::string    HttpRequest::getLineAndPopFromBuffer()
     return (line);
 }
 
-void    HttpRequest::printBuffer() const
+void HttpRequest::printBuffer() const
 {
     std::vector<unsigned char>::const_iterator it;
     for (it = _buffer.begin(); it != _buffer.end(); it++)
@@ -279,7 +160,7 @@ void    HttpRequest::printBuffer() const
     std::cout << std::endl;
 }
 
-void    HttpRequest::setClientSocket(int client_socket)
+void HttpRequest::setClientSocket(int client_socket)
 {
     _client_socket = client_socket;
 }
@@ -354,237 +235,6 @@ const std::map<std::string, std::string> &HttpRequest::getFormData() const
     return (_formData);
 }
 
-std::string HttpRequest::parseFieldname(const std::string &line, size_t *pos)
-{
-    if (!pos)
-        return (std::string());
-    size_t head_pos = *pos;
-    size_t end_pos = line.find(':', head_pos); 
-    if (end_pos == std::string::npos || end_pos <= head_pos) {
-        return (std::string()); 
-    }
-    std::string fieldname = line.substr(head_pos, end_pos - head_pos);
-    *pos = end_pos + 1;
-    return (fieldname); 
-}
-
-std::string HttpRequest::parseFieldValue(const std::string &line, size_t *pos)
-{
-    if (!pos)
-        return std::string();
-    while (isSpace(line[*pos]))
-        (*pos)++;
-    size_t start = *pos; 
-    size_t len = 0;
-    while (line[*pos + len] && !isSpace(line[*pos + len]) && line[*pos + len] != ';')
-        len++;
-    std::string fieldvalue = line.substr(start, len);
-    *pos += len;
-    while (isSpace(line[*pos]))
-        (*pos)++;
-    return (fieldvalue);
-}
-
-void HttpRequest::splitFormLine(const std::string &line, MultiFormData *form)
-{
-    //extract field name
-    size_t pos = 0;
-    std::string fieldname = parseFieldname(line, &pos);
-    if (fieldname.empty())
-    {
-        setStatusCode(BAD_REQUEST);
-        return ;   
-    }
-    //check if ':' is not skipped
-    if (line[pos] == ':')
-        pos++;
-    //check for OWS optional whitespace
-    while (isSpace(line[pos]))
-        pos++;
-    //extract field value
-    std::string fieldvalue = parseFieldValue(line, &pos);
-    if (fieldvalue.empty())
-    {
-        setStatusCode(BAD_REQUEST);
-        return ;   
-    }
-    //check extracted fieldvalue if there are more info
-    std::stringstream ss(line);
-    std::string token;
-    fieldname = toLower(fieldname);
-    while (std::getline(ss, token, ';'))
-    {
-        size_t equalPos = token.find('=');
-        if (equalPos != std::string::npos)
-        {
-            std::string key = token.substr(0, equalPos);
-            std::string value = token.substr(equalPos + 1);
-            
-            if (value[0] == '"' && value[value.size() - 1] == '"')
-                value = value.substr(1, value.size() - 2);
-            trimWhitespaces(key);
-            key = toLower(key);
-            if (!key.empty() && !value.empty())
-                form->fields[key] = value;
-        }
-    }
-    if (!fieldname.empty() && !fieldvalue.empty())
-        form->fields[fieldname] = fieldvalue;
-}
-
-
-void HttpRequest::parseUntilBinary(const std::string &boundary, MultiFormData *form)
-{
-    const std::string separator = "--" + boundary;
-
-    while (true)
-    {
-        std::string line = getLineAndPopFromBuffer();
-        
-        if (line == separator)
-            continue ;
-        if (line.empty())
-            break ;
-        splitFormLine(line, form);
-        if (_status != OK)
-            return ;
-    }
-    std::string type = form->fields["content-type"];
-    if (type != "image/jpeg" && type != "image/jpg" && type != "image/gif" && type != "image/png" && type != "image/bmp")
-    {
-        setStatusCode(UNSUPPORTED_MEDIA_TYPE);
-        return ;
-    }
-}
-
-void HttpRequest::parseBinary(const std::string &boundary, MultiFormData *form)
-{
-    if (_buffer.empty())
-    {
-        setStatusCode(INTERNAL_SERVER_ERROR);
-        return ;
-    }
-
-    std::string boundaryStr = "--" + boundary;
-    std::vector<unsigned char>::iterator boundaryIt;
-
-    while (true)
-    {
-        boundaryIt = std::search(_buffer.begin(), _buffer.end(), boundaryStr.begin(), boundaryStr.end());
-        
-        if (boundaryIt == _buffer.end())
-        {
-            form->binary.insert(form->binary.end(), _buffer.begin(), _buffer.end());
-            _buffer.clear();
-            break;
-        }
-
-        form->binary.insert(form->binary.end(), _buffer.begin(), boundaryIt);
-        _buffer.erase(_buffer.begin(), boundaryIt + boundaryStr.size());
-
-        if (std::string(_buffer.begin(), _buffer.begin() + 2) == "--")
-        {
-            _buffer.erase(_buffer.begin(), _buffer.begin() + 2);
-            break;
-        }
-    }
-}
-
-std::string byteToHex(unsigned char byte) {
-    std::stringstream ss;
-    ss << std::hex << std::setw(2) << std::setfill('0') << static_cast<int>(byte);
-    return ss.str();
-}
-
-void HttpRequest::parseMultipartForm(const std::string &boundary)
-{
-    if (_buffer.empty() || boundary.empty())
-    {
-        setStatusCode(INTERNAL_SERVER_ERROR);
-        return ;
-    }
-    MultiFormData form;
-    
-    while (!_buffer.empty())
-    {
-        parseUntilBinary(boundary, &form);
-        if (_status != OK)
-            return ;
-        parseBinary(boundary, &form);
-        if (_status != OK)
-            return ;
-        _multiFormData[form.fields["name"]] = form;
-    }
-    
-    // std::cout << "================================" << std::endl;
-    // std::cout << "Multipart Form Data" << std::endl;
-    // std::cout << "================================" << std::endl;
-    // std::cout << "_multiFormData size: " << _multiFormData.size() << std::endl;
-    // std::map<std::string, MultiFormData>::iterator it;
-    // for (it = _multiFormData.begin(); it != _multiFormData.end(); it++)
-    // {
-    //     std::cout << "MultiFormData key: " << it->first << std::endl;
-    //     std::cout << "fields size: " << it->second.fields.size() << std::endl;
-    //     std::map<std::string, std::string>::iterator field;
-    //     for (field = it->second.fields.begin(); field != it->second.fields.end(); field++)
-    //     {
-    //         std::cout << field->first << ": " << field->second << std::endl;
-    //     }
-    //     std::cout << "binary size: " << it->second.binary.size() << std::endl;
-    //     std::vector<unsigned char>::iterator binary;
-    //     for (binary = it->second.binary.begin(); binary != it->second.binary.end(); binary++)
-    //     {
-    //         std::cout << byteToHex(*binary) << " ";
-    //     }
-    //     std::cout << std::endl;
-    // }
-}
-
-void HttpRequest::parseRequestBody()
-{
-    std::string contentLen = getHeader("content-length");
-    if (contentLen.empty())
-    {
-        setStatusCode(BAD_REQUEST);
-        return ;
-    }
-    std::string type = getHeader("content-type");
-    if (type.empty())
-    {
-        setStatusCode(UNSUPPORTED_MEDIA_TYPE);
-        return ;
-    }
-    while (true)
-    {
-        if (_buffer.empty())
-            break;
-        std::string line = getLineAndPopFromBuffer();
-        if (line.empty())
-            continue ;
-        switch(_parseStep)
-        {
-            case REQUEST_BODY:
-            {
-                if (type == "application/x-www-form-urlencoded")
-                {
-                    parseFormData(line);
-                    if (_status != OK)
-                        return ;
-                }
-                std::string boundary;
-                if (isMultiPartFormData(&boundary)) 
-                {
-                    parseMultipartForm(boundary);
-                }
-                _parseStep = REQUEST_DONE;
-
-            }
-            default:
-                break ;
-        }
-    }
-}
-
 bool HttpRequest::isSupportedMediaPOST()
 {
     std::string type = getHeader("content-type");
@@ -610,29 +260,7 @@ bool HttpRequest::isSupportedMediaPOST(const std::string &type)
     return (false);
 }
 
-
-
-void HttpRequest::parseFormData(const std::string &line)
-{
-    std::stringstream ss(line);
-    std::string pair;
-
-    while (std::getline(ss, pair, '&'))
-    {
-        size_t equalPos = pair.find('=');
-        if (equalPos == std::string::npos)
-        {
-            setStatusCode(BAD_REQUEST);
-            return ;
-        }
-        std::string key = pair.substr(0, equalPos);
-        std::string value = pair.substr(equalPos + 1);
-        _formData[key] = value;
-    }
-}
-
-
-bool    HttpRequest::isMultiPartFormData(std::string *boundary)
+bool HttpRequest::isMultiPartFormData(std::string *boundary)
 {
     std::string type = getHeader("content-type");
     size_t pos = type.find("multipart/form-data");
@@ -645,7 +273,7 @@ bool    HttpRequest::isMultiPartFormData(std::string *boundary)
     return (true);
 }
 
-bool    HttpRequest::isMultiPartFormData()
+bool HttpRequest::isMultiPartFormData()
 {
     std::string type = getHeader("content-type");
     size_t pos = type.find("multipart/form-data");
@@ -654,7 +282,7 @@ bool    HttpRequest::isMultiPartFormData()
     return (true);
 }
 
-bool    HttpRequest::isForUpload()
+bool HttpRequest::isForUpload()
 {
     if (_multiFormData.empty())
         return (false);
