+++ srcs/http_response/Redirect.cpp
@@ -6,7 +6,7 @@
 /*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 01:02:29 by jdagoy            #+#    #+#             */
-/*   Updated: 2024/09/20 16:04:23 by jdagoy           ###   ########.fr       */
+/*   Updated: 2024/09/25 12:58:24 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
