+++ srcs/main.cpp
@@ -6,7 +6,7 @@
 /*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/05 00:23:30 by jdagoy            #+#    #+#             */
-/*   Updated: 2024/09/25 11:48:19 by jdagoy           ###   ########.fr       */
+/*   Updated: 2024/09/25 12:06:17 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
