+++ website/html/features.html
@@ -34,7 +34,7 @@
             <ul>
                 <li class="features-list"><a href="/old">Local redirect</a></li>
                 <li class="features-list"><a href="/campus19">External Redirect</a></li>
-                <li class="features-list"><a href="/old">Redirect to other port</a></li>
+                <!-- <li class="features-list"><a href="/old">Redirect to other port</a></li> -->
 
             </ul>
             <li>Errors</li>
@@ -51,7 +51,7 @@
 
         <h2>POST</h2>
         <ul>
-            <li class="features-list"><a href>Google Search</a></li>
+            <li class="features-list"><a href="/html/search.html">Google Search</a></li>
             <li class="features-list"><a href="/html/upload.html">Image Upload</a></li>
         </ul>
